requirements.txt
streamlit
pandas
altair
numpy
plotly.express
geopandas
